visitor board refac
